Bullet Hit Event
When the bullet hits an enemy, run an event
It just prints out the position where it hit an enemy for now

Was having trouble with the world and events so I reverted to the double
query collision system instead of checking the bullet's collisions for
enitities with an Enemy Component.

SystemStates are maybe the way to make this work, but I couldn't figure it
out.
This other way is simpler so that's probably better.