Shotgun
Added a shotgun. Fires 2 bullets in a spread

Shotgun Component
Not sure if this is the right pattern to have a gun component and a
shotgun component that modifies it. Doesn't seem especially scaleable for
multiple gun types.
fn shoot_bullet
Same issue with scaleability.
I'm not sure how to do the inheritance-type stuff I want to do in rust
yet.

Gauge
Want to do something when you hit with both bullets. (or hit with exactly
1 shot, but not both)
fn shotgun_both_hit is trying to track when both shots hit while trying to
account for both bullets being from the same volley/shot/trigger pull with
the shot_number.

It's unfinished.
It would probably work, but the problem is with resetting the vec storing
each shot's left and right bullets. I thought about resetting when you
reload, but will that have issues if bullets are still flying when you
finish reloading and reset? Probably

Version 2
New idea that might work better.