Shotgun Gauge
The idea is to have you gain some gauge when you hit with both shots in a
shotgun blast (or exactly 1, but not both)

The tracking now works. Got the first version to work so I didn't have to
do the alternate version 2. I thought I would have to reset the HitPair
vec on reload, but I actually need to reset the individual hitpair that
maps to this shot when the gun fires. This makes conflicts very unlikely.
It depends on fire rate, number of shots, reload time, and bullet life. If
you can cycle through your whole clip, reload, and shoot again within the
lieftime of the bullet, there could be conflicts where shot 0, gen 0
conflicts with shot 0 of gen 1(after reload).

HitPair vec is set up when the component is added with size == clip size.
Filled with None HitPairs. When you shoot, the hitpair is reset.

fn shotgun_check_gauge
checks if both shots in a shotgun volley have terminated and what the
outcome is. Just prints now, but could have more behaviour.
Reset the hitpair here too so it doesn't print forever.